+++ src/utils/timeMode.ts
@@ -1,0 +1,4 @@
+export default (time: string): 'exactly' | 'per' => {
+  if(time.includes(':')) return 'exactly';
+  else return 'per';
+};