+++ changeLog.md
@@ -1,2 +1,5 @@
 1. 添加手动触发功能
 2. 群聊添加黑/白名单模式
+3. 添加【每……分钟】发送模式
+
+本次更新不兼容过去版本的数据，请手动删除旧版数据文件。其位于`LiteLoaderQQNT数据目录/data/auto_send_messages/config.json`。